pyproject.toml
[tool.poetry]
name = "fastapi-cloud-observability"
version = "0.1.0"
description = ""
authors = ["Trayan Azarov <trayan.azarov@amikos.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
opentelemetry-api = "^1.22.0"
opentelemetry-exporter-otlp-proto-grpc = "^1.22.0"
opentelemetry-instrumentation-fastapi = "^0.43b0"
opentelemetry-sdk = "^1.22.0"
psutil = "^5.9.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
